% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{absolute_path_mapper}
\alias{absolute_path_mapper}
\title{Gets the absolute path of a file.}
\usage{
absolute_path_mapper(path)
}
\arguments{
\item{path}{a normalized, absolute path.}
}
\value{
The absolute host path, if it exists.
}
\description{
Gets the absolute path of a file.
}
