% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_age.R
\name{adjust_age}
\alias{adjust_age}
\title{Adjusting population age structure}
\usage{
adjust_age(data, location, year, AgeL, AgeH)
}
\description{
Smooth population counts using Arriaga and Sprague methods
}
