% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{argNames}
\alias{argNames}
\title{Get the names of the arguments to a function}
\usage{
argNames(x)
}
\arguments{
\item{x}{A function or string naming a function.}
}
\value{
A vector of the names of the arguments to a function.
}
\description{
Get the names of the arguments to a function
}
\examples{
argNames("readLines")
}
