% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{argumentNamedList}
\alias{argumentNamedList}
\title{Create a named list of length 1 using a name stored in a variable as the name.}
\usage{
argumentNamedList(name, x)
}
\arguments{
\item{name}{The name for the item in the list.}

\item{x}{The item to put in the list.}
}
\value{
A named list.
}
\description{
Create a named list of length 1 using a name stored in a variable as the name.
}
\examples{
argumentNamedList("test_name", 1)
}
