% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{bluefocus}
\alias{bluefocus}
\title{Blue as highlight}
\format{
An object of class \code{character} of length 8.
}
\usage{
bluefocus
}
\description{
Color scales for theme_simple
}
\keyword{datasets}
