% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{castDateString}
\alias{castDateString}
\title{Format a date string as "%Y-%m-%d" (YYYY-MM-DD), useful for converting dates selected
from a SQL database to a format compatible with a HTML date input value.}
\usage{
castDateString(x)
}
\arguments{
\item{x}{A string.}
}
\value{
A string, formatted YYYY-MM-DD.
}
\description{
Format a date string as "%Y-%m-%d" (YYYY-MM-DD), useful for converting dates selected
from a SQL database to a format compatible with a HTML date input value.
}
\examples{
castDateString(Sys.time())
}
