% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{castLogical}
\alias{castLogical}
\title{Convert strings to logical.}
\usage{
castLogical(x)
}
\arguments{
\item{x}{A string.}
}
\value{
A string, converted to logical.
}
\description{
Convert strings to logical.
}
\examples{
castLogical("1")
}
