% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{castNumeric}
\alias{castNumeric}
\title{Convert strings to numeric if possible, otherwise remains as is.}
\usage{
castNumeric(x)
}
\arguments{
\item{x}{A string.}
}
\value{
A string, converted to numeric if possible.
}
\description{
Convert strings to numeric if possible, otherwise remains as is.
}
\examples{
castNumeric("100")
}
