% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{consolidateList}
\alias{consolidateList}
\title{Group items of a list by name}
\usage{
consolidateList(x)
}
\arguments{
\item{x}{A named list, likely with names repeating for different positions.}
}
\value{
A list with items consolidated by name.
}
\description{
Group items of a list by name
}
\examples{
consolidateList(list("col1" = "Test", "col2" = "Hello", "col1" = "Repeated Name"))
}
