% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_db.R
\name{create_td_connection}
\alias{create_td_connection}
\title{Create a connection object to Teradata to then run queries with.}
\usage{
create_td_connection(
  td_id,
  password = NULL,
  jdbc_class_path = "~/terajdbc4.jar",
  url_td = NULL
)
}
\arguments{
\item{td_id}{string: Teradata ID}

\item{password}{string: Teradata password (only for Mac)}

\item{jdbc_class_path}{string: Path to terajdbc4.jar (only for Mac)}

\item{url_td:}{url to DB}
}
\value{
teradata connection object that is used to run queries
}
\description{
Developed for both windows and mac.
}
