% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugTransform.R
\name{debugTransform}
\alias{debugTransform}
\title{Debug Transform}
\usage{
debugTransform(dataset,features=attr(dataset,"features"))
}
\arguments{
\item{dataset, }{dataframe.}

\item{features, }{dataset's features list.}
}
\value{
Return the input dataset with 0.2 response rate on predicted variable.
}
\description{
It takes a dataset reded by \code{readDataset} and randomly assign 1's and 0's
to predicted variable according to \code{features} to reach 0.2 response rate.
}
\details{
This function is meant to be used for debugging, gives the chance to run the
\code{data_minning_template} with not so big datasets.
}
\author{
Daniel Fischer
}
