% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{detect_container_runtime}
\alias{detect_container_runtime}
\title{Check and Identify the Container Runtime}
\usage{
detect_container_runtime()
}
\value{
A list with two elements:
\item{name}{A character string indicating the container runtime: "Docker",
"Singularity", or "None".}
\item{is_running}{A logical value: \code{TRUE} if the process is running inside
a recognized container runtime, \code{FALSE} otherwise.}
}
\description{
The \code{detect_container_runtime} function checks if the current R process
is running  inside a Docker or Singularity container. It returns the name
of the container runtime (either "Docker", "Singularity", or "None") along
with a boolean indicating whether it is running inside any of these
containers.
}
\examples{
result <- detect_container_runtime()
cat("Container runtime:", result$name, "\n")
cat("Is running in container:", result$is_running, "\n")

}
