% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{do.call2}
\alias{do.call2}
\title{Filters the argument list to match the arguments in what and then calls do.call.}
\usage{
do.call2(what, args, quote = FALSE, envir = parent.frame())
}
\arguments{
\item{what}{See do.call.}

\item{args}{Argument list, gets filtered to match arguments of what. See do.call.}

\item{quote}{See do.call.}

\item{envir}{See do.call.}
}
\value{
See do.call.
}
\description{
Filters the argument list to match the arguments in what and then calls do.call.
}
\examples{
do.call2(intersect, list(x = c(1, 2, 3), y = c(2)))
}
\seealso{
do.call
}
