% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{docker_mount_mapper}
\alias{docker_mount_mapper}
\title{Maps a path to host volumes.}
\usage{
docker_mount_mapper(path)
}
\arguments{
\item{path}{a normalized, absolute path.}
}
\value{
The absolute host path, if it exists.
}
\description{
Maps a path to host volumes.
}
