% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryptCsv.R
\name{encryptCsv}
\alias{encryptCsv}
\title{Encrypt Csv}
\usage{
encryptCsv(input,
output,
encrypt.columns,
algorithm = "sha512",
batch.size = 1e5,
sep = ";",
...)
}
\arguments{
\item{input}{function input}

\item{output}{function output}

\item{encrypt.columns}{columns to encrypt}

\item{algorithm}{encryption algorithm}

\item{batch.size}{batch size}

\item{sep}{character separator}

\item{...}{arguments pass on methods.}
}
\description{
Encrypt Csv files
}
\details{
This function has no documentation yet.
}
