% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_function.R
\name{exam_function}
\alias{exam_function}
\title{Analisys and convertion of sparse matrix}
\usage{
exam_function(imput_matrix, output_file)
}
\arguments{
\item{imput_matrix}{Il percorso del file .mtx della matrice da leggere.}

\item{output_file}{Il nome o percorso del file .mtx in cui salvare la matrice sparsa.}
}
\value{
Nessun valore di ritorno esplicito. Stampa i risultati a schermo e scrive su disco.
}
\description{
this function read a dense matrix and convert it into a sparse one
it also calculate sparsity of zero elements e save on disk
}
