% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_deaths.R
\name{extract_deaths}
\alias{extract_deaths}
\title{Extract deaths from Demographic and Sanitary Statistics}
\usage{
extract_deaths(data, location, sex, year1, year2, omega = 100, trunc = 70)
}
\description{
Smooth death counts using PCLM method
}
