% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresScript.R
\name{featuresScript}
\alias{featuresScript}
\title{Features Script}
\usage{
featuresScript(dataset,features=attr(dataset,"features"))
}
\arguments{
\item{dataset}{dataset gotten by \code{readDataset()}.}

\item{features}{features list of the dataset.}
}
\value{
It return the dataset with the scripts executed for the variables specified in \code{features}.
}
\description{
It apply the function \code{transform(dataset,variable=script)} to each \code{variable} that has a \code{script} 
in the \code{features$script}.
}
\author{
Martin Vicencio
}
