% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstMonthDay.R
\name{firstMonthDay}
\alias{firstMonthDay}
\title{First Month Day}
\usage{
firstMonthDay(date)
}
\arguments{
\item{date}{A POSIXct, POSIXlt or Date parameter. Do NOT use character.}
}
\value{
The first day of the input's month as a Date
}
\description{
Recieve a date and return the first day of that date's month.
}
\author{
Daniel Fischer
}
