% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscper.R
\name{fiscper_month}
\alias{fiscper_month}
\alias{fiscper_date}
\alias{fiscper_quarter}
\alias{fiscper_semester}
\alias{fiscper_workdays}
\alias{make_fiscper}
\title{Fiscper functions}
\usage{
fiscper_month(fiscper)

fiscper_date(fiscper)

fiscper_quarter(fiscper)

fiscper_semester(fiscper)

fiscper_workdays(fiscper)

make_fiscper(year, month)
}
\arguments{
\item{fiscper}{Column containing the fiscper}

\item{year}{int of year}

\item{month}{int of month}
}
\value{

}
\description{
Imported from markfairbanks/graingr
}
\details{
Extract info from fiscper
}
