% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscper.R
\name{fiscper_year}
\alias{fiscper_year}
\title{Fiscper functions}
\usage{
fiscper_year(fiscper)
}
\arguments{
\item{fiscper}{Column containing the fiscper}
}
\value{

}
\description{
Imported from markfairbanks/graingr
}
\details{
Extract info from fiscper
}
