% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceClass.R
\name{forceClass}
\alias{forceClass}
\title{Force Classes}
\usage{
forceClass(dataset, features=attr(dataset, "features"))
}
\arguments{
\item{dataset}{dataset gotten by \code{readDataset()}}

\item{features}{features list. Takes the atribute from \code{dataset} by default.}
}
\value{
A dataset with the classes of the specified variables forced.
}
\description{
It forces the classes of certain variables according to the features list of the dataset.
}
\author{
Martin Vicencio
}
