% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_datatable.R
\name{format_datatable}
\alias{format_datatable}
\title{Creates a functional default datatable for interactive html outputs}
\usage{
format_datatable(
  data = NULL,
  caption = NULL,
  col_names = NULL,
  cols_show = NULL,
  ...
)
}
\arguments{
\item{data}{dataframe, tibble or tableone object with table content}

\item{caption}{datatable caption}

\item{col_names}{character vector of column names}

\item{cols_show}{character vector of columns to show (other are hidden but can be selected); if not specified, all are shown by default}

\item{...}{more parameters passed on}
}
\value{
returns a pretty table formatted for html or latex
}
\description{
Creates a functional default datatable for interactive html outputs
}
