% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gt.R
\name{format_gt}
\alias{format_gt}
\title{Automated \code{\link[gt:gt]{gt()}} Markdown/Quarto output for html or pdf}
\usage{
format_gt(tbl_gt = NULL, img_file_path = NULL, zoom = 1, dpi = 300)
}
\arguments{
\item{tbl_gt}{gt object}

\item{img_file_path}{file path where table images should be stored in pdf compilations; default is the current project directory}

\item{zoom}{A number specifying the zoom factor. A zoom factor of 2 will result in twice as many pixels vertically and horizontally. Note that using 2 is not exactly the same as taking a screenshot on a HiDPI (Retina) device: it is like increasing the zoom to 200 doubling the height and width of the browser window. This differs from using a HiDPI device because some web pages load different, higher-resolution images when they know they will be displayed on a HiDPI device (but using zoom will not report that there is a HiDPI device).}

\item{dpi}{DPI (dots per inch) value. Used to calculate the output width (in inches) of the images. This will be their actual width in pixels, divided by dpi. If not provided, the chunk option dpi is used; if NA, the output width will not be calculated.}
}
\value{
returns a pretty table formatted for html or latex
}
\description{
For complex tables, \code{\link[gt:gt]{gt()}} output formats are not optimized. This function
takes a finished \code{\link[gt:gt]{gt()}} table and outputs it either directly as an html table
or pdf table.

Note that in an interactive session no output will be displayed and the output is dependent on if the
Markdown/Quarto documented is being rendered towards html or pdf
}
