% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Creates pretty formatted tables with few lines of code for both html and pdfs}
\usage{
format_table(data = NULL, format_out = "html", font_size = 16, caption = NULL)
}
\arguments{
\item{data}{dataframe, tibble or tableone object with table content}

\item{format_out}{character, one of "html" or "latex" to choose from for optimized output}

\item{font_size}{numeric, font size of table}

\item{caption}{string, if table caption is desired}
}
\value{
returns a pretty table formatted for html or latex
}
\description{
Creates pretty formatted tables with few lines of code for both html and pdfs
}
