% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_structure.R
\name{get_age_structure}
\alias{get_age_structure}
\title{Estimate population counts from Census Statistics}
\usage{
get_age_structure(data, location, year, sex)
}
\description{
Smooth population counts using PCLM method
}
