% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mig_structure.R
\name{get_mig_structure}
\alias{get_mig_structure}
\title{Extract migration schedules (generalization of Tim Riffe's mig_un_fam)}
\usage{
get_mig_structure(
  NM,
  family,
  Single = TRUE,
  OAnew = 100,
  data = DemoTools::mig_un_families,
  params = DemoTools::mig_un_params
)
}
\description{
Smooth population counts using PCLM method
}
