% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure.R
\name{get_structure}
\alias{get_structure}
\title{Estimate population counts from Census Statistics}
\usage{
get_structure(data, area, age, year, sex, omega)
}
\description{
Smooth population counts using PCLM method
}
