% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{Get Features}
\usage{
getFeatures (features.file=paste0(dataset.path,"features.tab"), \cr
sep="\t")
}
\arguments{
\item{sep}{separator used in the features file}

\item{file}{path of the features file}
}
\value{
It returns the features list, wich contain relevant information to make tranformations to data.
}
\description{
It sets the features list from a file with the information.
}
\author{
Martin Vicencio
}
