% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResponseRate.R
\name{getResponseRate}
\alias{getResponseRate}
\title{Get Response Rate}
\usage{
getResponseRate(dataset, groups = unique(dataset$dataset_group),
  features = attr(dataset, "features"))
}
\arguments{
\item{dataset}{dataframe, dataset gotten by \code{readDatset}.}

\item{groups}{character vector, groups from which get response rates. Search on \code{dataset$dataset_group}}

\item{features}{list, features of dataset.}
}
\value{
It returns a two column dataframe with groups and their response rate.
}
\description{
Get the response rate of \code{features$predicted_var}, by groups.
}
\details{
This function is meant to be used as a complement for data mining fluxes that use \code{updateFeatures, getFeatures} function and others related.
}
\author{
Martin Vicencio
}
