% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScore.R
\name{getScore}
\alias{getScore}
\title{Get score}
\usage{
getScore(id_execution = NULL)
}
\arguments{
\item{id_execution}{integer vector indicating the id_execution of the scorings to get, 
if its \code{NULL} the it would get all \code{scr_score} table.}
}
\value{
return a list of the scorings named by \code{id_execution}, each scoring in the list has its \code{"metaData"}
 attribute gotten form \code{scr_metadata} table.
}
\description{
Get scoring from SQL data base in analytics server.
}
\author{
Martin Vicencio
}
