% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Inverse of the %in% function}
\usage{
... \%notin\% NA
}
\arguments{
\item{...}{See %in% documentation}
}
\value{
Boolean vector
}
\description{
Inverse of the %in% function
}
\examples{
a = 'a'
b = c('b','c','a')

a \%notin\% b
}
