% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{has_docker}
\alias{has_docker}
\title{Check if Docker is Available and Return Its Path}
\usage{
has_docker()
}
\value{
A list with two elements:
\item{found}{A logical value: \code{TRUE} if Docker is available in the system's
\code{PATH}, \code{FALSE} otherwise.}
\item{path}{A character string containing the full path to the Docker
executable if found,  or an empty string if not found.}
}
\description{
The \code{has_docker} function checks if the Docker executable is available
in the system's \code{PATH}. It returns a list with a boolean indicating if Docker
is found, and the path to the Docker executable if it is available.
}
\examples{
result <- has_docker()
if (result$found) {
  cat("Docker is available at:", result$path, "\n")
} else {
  cat("Docker is not available.\n")
}

}
