% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{has_singularity}
\alias{has_singularity}
\title{Check if Singularity is Available and Return Its Path}
\usage{
has_singularity()
}
\value{
A list with two elements:
\item{found}{A logical value: \code{TRUE} if Singularity is available in the
system's \code{PATH}, \code{FALSE} otherwise.}
\item{path}{A character string containing the full path to the Singularity
executable if found, or an empty string if not found.}
}
\description{
The \code{has_singularity} function checks if the Singularity executable is
available  in the system's \code{PATH}. It returns a list with a boolean
indicating if Singularity is found, and the path to the Singularity
executable if it is available.
}
\examples{
result <- has_singularity()
if (result$found) {
  cat("Singularity is available at:", result$path, "\n")
} else {
  cat("Singularity is not available.\n")
}

}
