% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{is_running_in_docker}
\alias{is_running_in_docker}
\title{Check if the script is running in Docker.}
\usage{
is_running_in_docker()

is_running_in_docker()
}
\value{
A truthy value indicating the state.

A truthy value indicating the state.
}
\description{
Check if the script is running in Docker.

Check if the script is running in a container.
}
