% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{is_running_in_singularity}
\alias{is_running_in_singularity}
\title{Check if the script is running in a container.}
\usage{
is_running_in_singularity()
}
\value{
A truthy value indicating the state.
}
\description{
Check if the script is running in a container.
}
