% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isContained.R
\name{isAreContained}
\alias{isAreContained}
\alias{isContained}
\alias{areContained}
\title{Is/Are Contained}
\usage{
isContained(polygon, x_0 = 0, y_0 = 0, point = NULL)

areContained(polygon, points)
}
\arguments{
\item{polygon}{matrix or dataframe of two columns. Ordered group of coordinated points that determinates a polygon.}

\item{x_0}{numeric, x coordenate of the point to be checked.}

\item{y_0}{numeric, y coordenate of the point to be checked.}

\item{point}{numeric vector, point given as c(x_0,y_0).}

\item{points}{dataframe or matrix, multplie points to ckeck with x coordenate on the first column and y coordenate on seccond.}
}
\value{
\code{boolean}, \code{TRUE} if the point its inside of the polygon, \code{FALSE} otherwise. If the point is rigth over a point or edge of the polygon it would return \code{FALSE}.
}
\description{
Cheks if a point is contained extrectely inside  of a polygon.
}
\examples{
   p=polygonBDA(edges=5,radius=2,alpha=0.4,color="green")
   
   ggplot()+p$plot 
   
   isContained(polygon=p$polygon,point=c(0,0))
   
   
   
   
   
   
   
   
   
   df=data.frame(x=runif(n = 30,min = -70.8,max =-70.5 ),y=runif(n = 30,min = -33.6,max =-33.3 ))

   ggplot()+geom_point(data=df,aes(x=x,y=y))

   circle=polygonBDA(x_0 = -70.65,y_0 = -33.45,edges = 360,radius = 0.1, alpha = 0.4 ,color = "grey")

   last_plot()+circle$plot

   df["InCircle"]=areContained(circle$polygon,df)

   ggplot()+circle$plot+geom_point(data=df,aes(x=x,y=y,colour=InCircle))
}
\author{
Martin Vicencio
}
