% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNull.R
\name{isNull}
\alias{isNull}
\title{Is Null}
\usage{
isNull(x,nulls)
}
\arguments{
\item{x}{array  with the elements to match. Don't use List nor Dataframes.}

\item{nulls}{array of "nulls" to be matched with. It takes NA, Inf, -Inf, NULL and NaN as "nulls" by default.}
}
\value{
A logical vector with True for the nulls elements' positions. Even if x is a matrix, the return value would be a vector.
}
\description{
Recieve an array and matches its elements with arbitrary "nulls".
}
\author{
Daniel Fischer
}
