% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{isNULLorNA}
\alias{isNULLorNA}
\title{Checks if x is NULL or NA}
\usage{
isNULLorNA(x)
}
\arguments{
\item{x}{A object.}
}
\value{
TRUE/FALSE.
}
\description{
Checks if x is NULL or NA
}
\examples{
isNULLorNA(NULL)
}
