% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpad.R
\name{lagpad}
\alias{lagpad}
\title{Lagpad}
\usage{
lagpad(x,k,default)
}
\arguments{
\item{x}{array to be displaced.}

\item{k}{the number of positions to displace. If its negative it would displace from the end.}

\item{default}{value to fill the displaced empty spots. It takes 0 by default.}
}
\value{
Returns an array of same length as x, with x's values displaced k positions and the default values where its supposed to.
}
\description{
Displaces an array the first or last k positions and fill the dislpaced empty spots with a specific value.
}
\author{
Daniel Fischer
}
