% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastMonthDay.R
\name{lastMonthDay}
\alias{lastMonthDay}
\title{Last  Month Day}
\usage{
lastMonthDay(date)
}
\arguments{
\item{date}{A POSIXct, POSIXlt or Date parameter. Do NOT use character.}
}
\value{
The Last  day of the input's month as a Date.
}
\description{
Recieve a date and return the Last  day of that date's month.
}
\author{
Daniel Fischer
}
