% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftCurve.R
\name{liftCurve}
\alias{liftCurve}
\title{Lift Curve}
\usage{
liftCurve(score, quantiles=100, cumulative=T, groupCol=NULL)
}
\arguments{
\item{score}{List of scores as dataframes or a single dataframe with al scores binded.}

\item{quantiles}{integer, number of buckets to make the lift curve.}

\item{cumulative}{boolean, Is cumulative lift?}

\item{groupCol}{character indicating the name of the column to group-by the differents 
score groups if \code{score} is a single dataframe.}
}
\value{
It returns a three column dataframe \code{quantile}, \code{lift}, and another one to indentify each group of scoring.
}
\description{
Get the Lift curve(s) from a(some) score(s) dataframe(s)
}
\author{
Martin Vicencio
}
