% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limitFactors.R
\name{limitFactors}
\alias{limitFactors}
\title{Limit Factor}
\usage{
limitFactors(dataset,features= attr(dataset,"features"), newlevel= "others")
}
\arguments{
\item{dataset}{dataset with the factor variables.}

\item{features}{features list of dataset, or of a previous dataset to replicate its transformations.}

\item{newlevel}{character, name of the new level created.}
}
\value{
Return the input dataset with the factor columns modificated according to the features list.
}
\description{
It reduce the levels of a factor variable according to \code{features$factor_limit} preserving the most frequent levels and gathering the rest into a new level called by \code{newlevel}.
Once that the levels were reduced it adds a new element to dataset's features called \code{factors}, 
wich is a list of character vectors, each vector is named by a factor variable of the output dataset and contain the levels of that variable. This new element is usefull to replicate the reductions
made for the actual dataset to others datasets in the future.
}
\details{
For each variable it take the limit from \code{features$factor_limit}. If the limit is n
 it would preserve the n most frequents, and generate a new one called \code{newlevel}, unless that 
 there were only n+1 levels. Anyway it would end with n+1 levels.\cr
 \cr
 In case that this function is used to replicate the transformations made to a previus dataset using its features, 
 it will preserve the levels according to \code{features$factors} for all factors variables, and assing any new 
 level of a variable  to the mos frequent level of that variable.
}
\author{
Martin Vicencio
}
