% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{listExtract}
\alias{listExtract}
\title{Extract the values from each entry in a list of vectors at a specific index}
\usage{
listExtract(x, pos)
}
\arguments{
\item{x}{A list, each item of the list should have equal length.}

\item{pos}{A integer, the position to extract from each entry in the list.}
}
\value{
A list.
}
\description{
Extract the values from each entry in a list of vectors at a specific index
}
\examples{
listExtract(list(col1 = c(1, 2, 3, 4, 5), col2 = c("a", "b", "c", "d", "e")), 3)
}
