% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\name{loadModel}
\alias{loadModel}
\title{Load Model}
\usage{
loadModel(file)
}
\arguments{
\item{file, }{faile path of the calibration object.}
}
\value{
The calibration Object allocated in \code{file}
}
\description{
Load calibration object of a model with \code{readRDs}
}
