% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataFrameDistributionConversor.R
\name{makeDataFrameDistributionConversor}
\alias{makeDataFrameDistributionConversor}
\title{Distribution conversor for data frames}
\usage{
makeDataFrameDistributionConversor(dataset,\cr 
columns = colnames(dataset),\cr
objectiveDistribution = qnorm,\cr
...)
}
\arguments{
\item{dataset}{dataframe with the columns of data samples to make a distribution of.}

\item{columns}{character vector with the names of the columns to make a distribution of.}

\item{objectiveDistribution}{Quantile function of the distribution to convert in (qnorm,qexp,qunif,...)}

\item{...}{parameters to pass to \code{objectiveDistribution}.}
}
\value{
Returns a function that recieve a data frame and returns a data frame with the selected column converted to columns that share
the same cumulative probability in the objective distribution for every row.
}
\description{
Make distribution from columns of a data frame and passes every one to other 
 distribution by equaling the probabilities of a quantile.
}
\details{
This function bassically apply \code{makeDistributionConversor()} to selected columns.
}
\seealso{
\code{makeDistributionConversor} documentation from this package.
}
\author{
Daniel Fischer
}
