% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeScale.R
\name{makeScale}
\alias{makeScale}
\title{Make Scale}
\usage{
makeScale(table, cols= colnames(table), mean.function=mean, sd.function=sd)
}
\arguments{
\item{table}{table that contain the data.}

\item{cols}{character vector with names of columns to normalize.}

\item{mean.function}{function to get the mean value to use in Z score.}

\item{sd.function}{function to get the standard deviation value to use in Z score.}
}
\value{
Returns a list that contain 4 elements, the functions scale(table) and unScale(table) which recieve a table an returns
a normalized or un-normalized one, a vector with the mean value of each column with its name and another vector with
the standard deviation value of each column with its name.
}
\description{
Normalize tables by columns with Z score. This function returns the mean value and standard deviation
of each column, and the functions to pass from the original table to a normilized one and vice versa.
}
\author{
Daniel Fischer
}
