% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow.R
\name{mlflow_log_plot}
\alias{mlflow_log_plot}
\title{Save plot to MLFlow}
\usage{
mlflow_log_plot(plot, plot_path, height = 9, width = 5)
}
\arguments{
\item{plot}{object GGplot object}

\item{plot_path}{string Local path to save plot file to}

\item{height}{int Height of plot}

\item{width}{int Width of plot}
}
\value{

}
\description{
Save plot to MLFlow
}
