% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow.R
\name{mlflow_log_table}
\alias{mlflow_log_table}
\title{Save dataframe to a stargazer/.txt table and MLFlow}
\usage{
mlflow_log_table(table, table_path, title = "Table Name")
}
\arguments{
\item{table}{dataframe/tibble Dataframe to save as a .txt table}

\item{table_path}{string Local path to save the table (.txt) to}

\item{title}{string Title of table}
}
\value{

}
\description{
Save dataframe to a stargazer/.txt table and MLFlow
}
