% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{modifier_timed_directory}
\alias{modifier_timed_directory}
\title{Append a timed directory with the format \%Y\%m\%d_\%H\%M\%S to the
target_directory path.}
\usage{
modifier_timed_directory(target_directory)
}
\arguments{
\item{target_directory}{The target directory.}
}
\value{
Timed directory with the format \%Y\%m\%d_\%H\%M\%S appended to
target_directory path.
}
\description{
Append a timed directory with the format \%Y\%m\%d_\%H\%M\%S to the
target_directory path.
}
