% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{normalize_path}
\alias{normalize_path}
\title{Normalizes a path}
\usage{
normalize_path(path, path_mappers = c())
}
\arguments{
\item{path}{The path to normalize.}

\item{path_mappers}{The mappers to be utilized to normalize the path.}
}
\value{
The normalized path.
}
\description{
Normalizes a path
}
