% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericMonth.R
\name{numericMonth}
\alias{numericMonth}
\title{Numeric Month}
\usage{
numericMonth(months)
}
\arguments{
\item{months}{array of months to be tunred. Every element of months has to be a Character or able
to be coerced into one.}
}
\value{
Returns an array of same length as months, with numeric representation of its months.
}
\description{
Takes a month, or array of months, wrote and turns it into numeric representation.
}
\details{
It accepts spanish months too.
}
\author{
Daniel Fischer
}
