% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{pastePaths}
\alias{pastePaths}
\title{Paste parts of file paths/urls separated with single forward-slashes}
\usage{
pastePaths(...)
}
\arguments{
\item{...}{Text strings to combine into a file path}
}
\value{
A string.
}
\description{
Paste parts of file paths/urls separated with single forward-slashes
}
\examples{
pastePaths("/home/", "/files")
}
