% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonLikeStringConcat.R
\name{+}
\alias{+}
\title{Python like string concatenation}
\usage{
"+"(e1, e2)
}
\arguments{
\item{e1}{Character vector}

\item{e2}{Character vector}
}
\value{
Character vector
}
\description{
This is a function that allow you to do string concation the python way.
It behave juste like the paste0 function.
}
