% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonBDA.R
\name{polygonBDA}
\alias{polygonBDA}
\title{Polygon BDA}
\usage{
polygonBDA(x_0=0,y_0=0,edges=3,radius=1,alpha=0.3)
}
\arguments{
\item{x_0}{numeric, x coordenate of the polygon's center.}

\item{y_0}{numeric, y coordenate of the polygon's center.}

\item{edges}{integer, number of edges of the polygon.}

\item{radius}{numeric, polygon's radius.}

\item{alpha}{numeric between 0-1 indicating the opacity of the polygon.}
}
\value{
It returns a list with two elements. \code{\$polygon} is a dataframe with the points of the polygon and \code{\$plot} is a ggplot obect to be add to a ggplot.
}
\description{
Draw and get a regular polygon.
}
\details{
To plot this polygon do something like \code{ggplot()+polygonBDA(...)$plot}.
}
\examples{
   p=polygonBDA(edges=5,radius=2,alpha=0.4,color="green")
   
   p$polygon
   
   ggplot()+p$plot    
}
\author{
Martin Vicencio
}
