% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictBDA.R
\name{predictBDA}
\alias{predictBDA}
\title{Predict BDA}
\usage{
predictBDA(calibration,dataset,subset = unique(dataset$dataset_group),features=attr(dataset,"features"))
}
\arguments{
\item{calibration}{a calibration object of a model}

\item{dataset}{dataframe with data to be scored}

\item{subset}{character vector indicating the groups of data to be scored, it groups by \code{dataset$dataset_group}, by default take all groups.}

\item{features}{features list of dataset.}

\item{...}{arguments pass on \code{predict()} methods.}
}
\value{
a list of socrings with their attributes and dataframes named by the groups of subsetting.
}
\description{
Score datasets according a calibrated model
}
\details{
First, it apply \code{predict} function with \code{calibration} and data subsetted from \code{dataset}, 
after it transform the score dataframe to keep the same format for all calibration classes. It Also pass \code{features}
 and some relevant attributes of the prediction to \code{JSON} to make a new attribute called \code{metaData}. 
 This new attribute is a dataframe with information of the execution of the scoring like type of model, date of execution
  and name of project. This attribute is added to upload the scoring results and information to analytics server with \code{saveScore}
   function.
}
