% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictKmeans.R
\name{predictKmeans}
\alias{predictKmeans}
\title{Predict K-means}
\usage{
predictKmeans(fit,dataset,escalator = NULL, segment.names = NULL)
}
\arguments{
\item{fit}{fitting}

\item{dataset}{Data array}

\item{escalator}{escalator, \code{NULL} by default}

\item{segment.names}{segments' names}
}
\description{
Predict with k-means method.
}
\details{
This function has no documentation yet
}
