% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_df.R
\name{prep_df}
\alias{prep_df}
\title{Prepare Data Frame}
\usage{
prep_df(x, standard.names = TRUE, standard.char = TRUE, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{standard.names}{Should column names be standardized?, Default: TRUE}

\item{standard.char}{Should character column strings be standardized?, Default: TRUE}
}
\value{
A standardized data frame.
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 data("starwars")
 test.df <- head(starwars)
 prep_df(test.df)
 prep_df(test.df, replacement.char = "+")
 }
}
}
