% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{print_html_as_png}
\alias{print_html_as_png}
\title{Print a widget image to png}
\usage{
print_html_as_png(x, file, background = "white")
}
\arguments{
\item{x}{a browsable htmlwidget}

\item{file}{a file path to write the image to}

\item{background}{colour for the image}
}
\value{
invisibly the file path of written image
}
\description{
Prints a browsable htmlwidget (that for example might show
in RStudio viewer pane to a png file)
}
