% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectFile.R
\name{projectFile}
\alias{projectFile}
\title{Project File}
\usage{
projectFile(filename)
}
\arguments{
\item{filename}{character with the name of the file}
}
\value{
the path storage as \code{project.files.path} with "\code{/filename}" appended.
}
\description{
Append the name of a file to \code{project.file.path}.
}
\details{
\code{project.files.path} must be previously declarated.
}
