% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{qualitative1}
\alias{qualitative1}
\title{Qualitative color schemes by Paul Tol}
\format{
An object of class \code{character} of length 1.
}
\usage{
qualitative1
}
\description{
Color scales for theme_simple
}
\keyword{datasets}
