% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{qualitative10}
\alias{qualitative10}
\title{Qualitative color schemes by Paul Tol}
\format{
An object of class \code{character} of length 10.
}
\usage{
qualitative10
}
\description{
Color scales for theme_simple
}
\keyword{datasets}
