% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{qualitative11}
\alias{qualitative11}
\title{Qualitative color schemes by Paul Tol}
\format{
An object of class \code{character} of length 11.
}
\usage{
qualitative11
}
\description{
Qualitative color schemes by Paul Tol
}
\keyword{datasets}
