% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{qualitative2}
\alias{qualitative2}
\title{Qualitative color schemes by Paul Tol}
\format{
An object of class \code{character} of length 2.
}
\usage{
qualitative2
}
\description{
Qualitative color schemes by Paul Tol
}
\keyword{datasets}
