% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{qualitative4}
\alias{qualitative4}
\title{Qualitative color schemes by Paul Tol}
\format{
An object of class \code{character} of length 4.
}
\usage{
qualitative4
}
\description{
Qualitative color schemes by Paul Tol
}
\keyword{datasets}
