% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{qualitative6}
\alias{qualitative6}
\title{Qualitative color schemes by Paul Tol}
\format{
An object of class \code{character} of length 6.
}
\usage{
qualitative6
}
\description{
Color scales for theme_simple
}
\keyword{datasets}
