% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{qualitative7}
\alias{qualitative7}
\title{Qualitative color schemes by Paul Tol}
\format{
An object of class \code{character} of length 7.
}
\usage{
qualitative7
}
\description{
Color scales for theme_simple
}
\keyword{datasets}
