% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_db.R
\name{read_sql_file}
\alias{read_sql_file}
\title{Read in SQL file as a string}
\usage{
read_sql_file(filepath)
}
\arguments{
\item{filepath}{Path to SQL file}
}
\value{
string format of the SQL file
}
\description{
Read in SQL file as a string
}
