% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDataset.R
\name{readDataset}
\alias{readDataset}
\title{Read Datasets}
\usage{
readDataset (files,\cr
features.list,\cr
sep = ";",\cr
  x,\cr
...)
}
\arguments{
\item{files}{character vector with files' names}

\item{features.list}{feature list gotten by \code{getFeatures()}}

\item{sep}{separator used in files}

\item{...}{arguments to pass on fread method like \code{nrows}.}
}
\value{
A data frame made of all files in \code{files}. with \code{features.list} added as a attribute called "features"
}
\description{
Take a vector with files, read them according a features list and combine them by columns filling the empty columns with \code{NA\'s}
}
\details{
It only reads the \code{features.list$allSelected} variables. And it adds one column to the returned dataset
that correspond to the origin file of the row.
}
