% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance.R
\name{rebalance}
\alias{rebalance}
\title{Rebalance datasets}
\usage{
rebalance(dataset,subset,responseRate,features=attr(dataset,"features"))
}
\arguments{
\item{dataset, }{dataset gotten by \code{readDataset()}.}

\item{subset, }{name off the subset to rebalance. It search on \code{dataset$dataset_group}.}

\item{responseRate, }{numeric between 0 and 1 indicating the objetive response rate for predicted variable.}

\item{features, }{features list of dataset. It defines the \code{features$predicted_var} for rebalancement.}
}
\value{
It returns a identical dataset as input but with less rows and the subset specified re-balanced.
}
\description{
It takes a dataset and rebalance a specific subset gotten by \code{readDataset()}. 
It randomly deletes the minimum amount of rows from a dataset to reach the especified responseRate.
}
\details{
If you want the \code{"train"} subset to be balance with 20% of response rate on \code{predicted_var}
you will use \code{rebalance(dataset,"train",0.2)}. And this function will delete rows of the subset without 
preferences to reach the \code{responseRate}. This function depends off \code{getFeatures() and readDataset()} do not edit separetly.
}
\author{
Martin Vicencio
}
