% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_non_alphanumeric_char.R
\name{replace_non_alphanumeric_char}
\alias{replace_non_alphanumeric_char}
\title{Replace Non-Alphanumeric Characters}
\usage{
replace_non_alphanumeric_char(x, replacement.char = "_")
}
\arguments{
\item{x}{A characte vector.}

\item{replacement.char}{The character(s) used to replace all of the non-alphanumeric characters., Default: '_'}
}
\value{
A vector
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
test.vec <- c("blah&*\%^#blah", "blah*\%   ()blah", "1234+==-:;56789")
 replace_non_alphanumeric_char(test.vec)
 replace_non_alphanumeric_char(test.vec, replacement.char = "+")
 }
}
}
