% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceNull.R
\name{replaceNull}
\alias{replaceNull}
\title{Replace Null}
\usage{
replaceNull(x,replacement,...)
}
\arguments{
\item{x}{array. Don't use List nor Dataframes.}

\item{replacement}{value or array to replace nulls.}

\item{...}{array of "nulls". It takes NA, Inf, -Inf, NULL and NaN as "nulls" by default.}
}
\value{
Returns an array with nulls replaced.
}
\description{
Recieve an array and replaces the "nulls" elements of it by a replacement.
}
\details{
If replacement is an array with length L shorter than x's length, the funciton would replace the null on the position
K*L+i of x, by replacement's i-element. This means the function would go through both arrays at same, repeating the replacement till the end
}
\author{
Daniel Fischer
}
