% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceNullFeatures.R
\name{replaceNullFeatures}
\alias{replaceNullFeatures}
\title{Replace Nulls by features.}
\usage{
replaceNullFeatures(dataset,features= attr(dataset,"features"))
}
\arguments{
\item{dataset}{dataset with the variables to be replaced.}

\item{features}{features list of dataset.}
}
\value{
Return the input dataset with the "nulls" values changed by the features' replacement.
}
\description{
It replaces the \code{NA,NULL,inf,-inf,NaN} values of a variable according to the features list.
}
\details{
This function calls \code{replaceNull()} from \code{toolkitEntel} package. It replace the "nulls" with the characters
  that are indicated in \code{features$replace_na}.
}
\author{
Martin Vicencio
}
