% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurve.R
\name{rocCurve}
\alias{rocCurve}
\title{ROC Curve}
\usage{
rocCurve(score,groupCol=NULL)
}
\arguments{
\item{score}{List of scores as dataframes or a single dataframe with al scores binded.}

\item{groupCol}{character indicating the name of the column to group-by the differents 
score groups if \code{score} is a single dataframe.}
}
\value{
It returns a three column dataframe:\cr
\code{FPR} false positive rate, x-axis.\cr
\code{TPR} true positive rate, y-axis\cr
 And another one to indentify each group of scoring.
}
\description{
Get the ROC curve(s) from a(some) score(s) dataframe(s)
}
\author{
Martin Vicencio
}
