% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocPlot.R
\name{rocPlot}
\alias{rocPlot}
\title{ROC Plot}
\usage{
rocPlot(score,aucDigits=4,title="ROC",legendTitle=NULL,xlab=NULL,ylab=NULL,showAuc=T,groupCol=NULL)
}
\arguments{
\item{score}{List of scores as dataframes or a single dataframe with al scores binded.}

\item{aucDigits}{integer indicating the number of decimal places to be used for auc indicators.}

\item{title, legendTitle, xlab, ylab}{characters, plot tittle, legend title, x-axis label and y-axis label.}

\item{showAuc}{boolean, plot the AUC of each roc curve?}

\item{groupCol}{character indicating the name of the column to group-by the differents 
score groups if \code{score} is a single dataframe.}
}
\value{
It returns a ggplot object so it can be post modificated.
}
\description{
Plot the  ROC curve(s) from a(some) score(s) dataframe(s)
}
\author{
Martin Vicencio
}
