% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{run_in_container}
\alias{run_in_container}
\title{Run a container.}
\usage{
run_in_container(image_name, volumes = list(), additional_arguments = c())
}
\arguments{
\item{image_name}{The image you want to run.}

\item{volumes}{The list of volumes to mount to the container.}

\item{additional_arguments}{Vector of arguments to pass to the container.}
}
\description{
Run a container.
}
