% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run_in_docker}
\alias{run_in_docker}
\title{Run a docker container.}
\usage{
run_in_docker(image_name, volumes = list(), additional_arguments = c())
}
\arguments{
\item{image_name}{The docker image you want to run.}

\item{volumes}{The list of volumes to mount to the container.}

\item{additional_arguments}{Vector of arguments to pass to the container.}
}
\description{
Run a docker container.
}
