% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{run_in_singularity}
\alias{run_in_singularity}
\title{Run a singularity container.}
\usage{
run_in_singularity(image_name, volumes = list(), additional_arguments = c())
}
\arguments{
\item{image_name}{The singularity image you want to run.}

\item{volumes}{The list of volumes to mount to the container.}

\item{additional_arguments}{Vector of arguments to pass to the container.}
}
\description{
Run a singularity container.
}
