% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_db.R
\name{run_query}
\alias{run_query}
\title{Run queries with teradata connection}
\usage{
run_query(td_con, sql_query)
}
\arguments{
\item{td_con}{object: TD connection object from create_teradata_connection()}

\item{sql_query}{string: SQL query to run. Can be multi-line.}
}
\value{
tibble: Dataframe of SQL results
}
\description{
Run queries with teradata connection
}
