% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeRbind.R
\name{safeRbind}
\alias{safeRbind}
\title{Safe Rbind}
\usage{
safeRbind(dataframes,\cr
cols = unique(do.call(c, llply(dataframes, colnames))))
}
\arguments{
\item{dataframes}{a list of dataframes.}

\item{cols}{the columns to keep after combination, by default it takes all the columns of every dataframe}
}
\value{
Return a unique dataframe made of all dataframes in \code{dataframes}, they appends in order, so the first row 
belongs to the first dataframe, and the last to the last.
}
\description{
Recieve a list of data tables or data frames and combine them into one by columns. It fills the missing columns in a dataframe with \code{NA\'s}
}
\author{
Martin Vicencio
}
