% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{sampleStr}
\alias{sampleStr}
\title{Generates (pseudo)random strings of the specified char length}
\usage{
sampleStr(n_char, sample_chars = c(letters, LETTERS, 0:9))
}
\arguments{
\item{n_char}{A integer, the number of chars to include in the output string.}

\item{sample_chars}{A vector of characters to sample from. Includes the lowercase and uppercase English alphabet and 0-9 by default.}
}
\value{
A string.
}
\description{
Generates (pseudo)random strings of the specified char length
}
\examples{
sampleStr(10)
}
