% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModel.R
\name{saveModel}
\alias{saveModel}
\title{Save Model}
\usage{
saveModel(calibration,file,features=attr(calibration,"features"))
}
\arguments{
\item{calibration}{a calibrated model.}

\item{file}{character, path of file that would be storaged.}

\item{features}{features list Takes the atribute from \code{calibration} by default.}
}
\description{
It saves a calibrated model with a features list as an attribute of it.
}
\details{
This function is meant to be used on models gotten by \code{trainModel}
}
