% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveScore.R
\name{saveScore}
\alias{saveScore}
\title{Save Score}
\usage{
saveScore(score, description = paste("score ejecutado en:",as.character(Sys.Date())))
}
\arguments{
\item{score}{list, socoring gotten by \code{predictBDA}. Or a dataframe.}

\item{description}{character, a brief description of the scoring to be storage in the server.}
}
\description{
Save scoring to SQL database in analytics server.
}
\details{
This function upload data in two tables from analytics, \code{scr_metadata} and \code{scr_score}. 
The first correspond to the \code{metaData} attribute, inherit from \code{predictBDA}, of the scoring, and the 
sccond to the dataframes in \code{score}.
}
\author{
Martin Vicencio
}
