% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simple.R
\name{scale_dollar_labels}
\alias{scale_dollar_labels}
\title{Currency}
\usage{
scale_dollar_labels(labels)
}
\arguments{
\item{labels}{vector of labels}
}
\value{
formatted labels
}
\description{
Currency
}
