% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simple.R
\name{scale_percent_labels}
\alias{scale_percent_labels}
\title{Percentages}
\usage{
scale_percent_labels(labels)
}
\arguments{
\item{labels}{vector of labels}
}
\value{
formatted labels
}
\description{
Percentages
}
