% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_x_longitude}
\alias{scale_x_longitude}
\title{Longitude}
\usage{
scale_x_longitude(xmin = -180, xmax = 180, step = 1, ...)
}
\arguments{
\item{labels}{vector of labels}
}
\value{
formatted labels
}
\description{
Longitude
}
