% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_y_latitude}
\alias{scale_y_latitude}
\title{Latitude}
\usage{
scale_y_latitude(ymin = -90, ymax = 90, step = 0.5, ...)
}
\arguments{
\item{labels}{vector of labels}
}
\value{
formatted labels
}
\description{
Latitude
}
