% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scan}
\alias{scan}
\title{SAS like scan function}
\usage{
scan(x, split, n, ...)
}
\arguments{
\item{x}{Character vector}

\item{split}{The splitting character}

\item{n}{Position of the substring to extract}

\item{...}{Arguments to pass to the strsplit function}
}
\value{
Character vector
}
\description{
SAS like scan function
}
\examples{
scan(x = 'ab de fg',split = ' ',n = 2)
}
