% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{simple_okabeito}
\alias{simple_okabeito}
\title{Okabe-Ito}
\format{
An object of class \code{character} of length 15.
}
\usage{
simple_okabeito
}
\description{
Color scales for theme_simple
}
\keyword{datasets}
