% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skeleton}
\alias{skeleton}
\title{skeleton}
\usage{
skeleton(
  input_dir_path,
  countmatrix_name,
  metadata_name,
  reference_group,
  organism
)
}
\arguments{
\item{input_dir_path, }{a character string indicating the path of the directory containing the fastq files and the csv files obtained from the indexing.Important this must be always present, the other parameters can be modified}

\item{countmatrix_name, }{name of the count matrix file obtained after the genome indexing}

\item{metadata_name, }{name of the metadata file obtained after the genome indexing}

\item{reference_group, }{name of the reference group inside the meetadata (wt, cpes,...)}

\item{organism, }{name of the organism subject of the analysis. Supported organisms are: 'Homo sapiens', 'Mus musculus' or 'Drosophila melanogaster'}
}
\description{
This function is the basic for designing a local front end for function based on Rscript embedded in docker containers.
}
\examples{
\dontrun{
skeleton(
  input_dir_path = "/the/input/dir",
  countmatrix_name = "gene_count_matrix.csv",
  metadata_name = "Covariatesstat.csv",
  reference_group = "wt",
  organism = "Drosophilamelanogaster"
)
}
}
\author{
Luca Alessandri, Agata D'Onofrio, Eliseo Martelli
}
