% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowCluster.R
\name{snowCluster}
\alias{snowCluster}
\title{Snow Cluster}
\usage{
snowCluster(n = parallel::detectCores(),scripts = c("./init.R"))
}
\arguments{
\item{n}{number of workers}

\item{scripts}{R scripts to execute in each worker}
}
\description{
This function has no documentation yet
}
\details{
this function execute this code:\cr
 \code{snowCluster = function(n = parallel::detectCores(),scripts = c("./init.R")){\cr
safeLibrary("doSNOW")\cr
cl = makeSOCKcluster(n,outfile="cl.txt")\cr
registerDoSNOW(cl)\cr
lapply(scripts, function(x) clusterCall(cl,source,x))\cr
return(cl)\cr
}}
}
