% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceDir.R
\name{sourceDir}
\alias{sourceDir}
\title{Source Directory.}
\usage{
sourceDir(path, encoding= "UTF-8", exclude= c(), ...)
}
\arguments{
\item{path}{character with the directory's path from we want to read R files.}

\item{encoding}{character vector. The encoding(s) to be assumed.}

\item{exclude}{character vector. Files we don't want to be readed as "filename.R".}

\item{...}{All other parameters that are used in source function.}
}
\description{
Reads R codes from a directory with its path. This is a directory version of source function from base.
}
\seealso{
Source function's dodumentation for more details, using help(source).
}
\author{
Daniel Fischer
}
