% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderBDA.R
\name{spiderBDA}
\alias{spiderBDA}
\title{Spider charts BDA}
\usage{
spiderBDA(centroids, show.max = T, show.values = T, show.axis = F,
  color = "orange", label.digits = 3, label.size = 3, nudge.x = 0.25,
  nudge.y = 0.05, legend.position = c(0.75, 0.05), legend.text.size = 8,
  many.vars = F, multiplot = F, multiplot.file = "", multiplot.cols = 2,
  multiplot.layout = NULL)
}
\arguments{
\item{centroids}{dataframe, centroids to plot as rows and variables as columns.}

\item{show.max}{boolean, show the maximum vlaues of each variable in the plot?}

\item{show.values}{boolean, show the centroid's value of each varaible in the plot?}

\item{show.axis}{boolean, show X and Y axis in plots?}

\item{color}{character, color of the chart like \code{"orange"} or \code{"#234567"}.}

\item{label.digits}{integer, number of digits after point for labels with the values.}

\item{label.size}{numeric, label size.}

\item{nudge.x}{numeric, nudge for the labels in X-axis.}

\item{nudge.y}{numeric > 0, nudge for the y-axis. The maximum's and centroid's labels moves in opposite directions to not overlap themself.}

\item{legend.position}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector).}

\item{legend.text.size}{numeric, legend text size.}

\item{many.vars}{boolean, if \code{TRUE}, then it plot with diferents shapes to make it easier to identify each variable.}

\item{multiplot}{boolean, print all plots together ?}

\item{multiplot.file}{character, path where storage the multiplot in PDF like "../plots/multiplot.pdf". By default it takes \code{""}, so it doesn't storage anything.}

\item{multiplot.cols}{integer, number of columns to make the multiplot}

\item{multiplot.layout}{matrix, custom layout fot the multiplot.}
}
\value{
It return a list of ggplots.
}
\description{
Make awesome spider charts
}
\details{
IT is recomended to try multiple combination of parameters before storage any multiplot to get better plots.
}
\examples{
   centroids=data.frame(rnorm(9),rnorm(9),rnorm(9),rnorm(9),rnorm(9),rnorm(9),rnorm(9),rnorm(9))
   names(centroids)=letters[1:8]
   a=spiderBDA(centroids,show.max=F,show.values=F,nudge.x=0.1,nudge.y=0.1,legend.position = "bottom",legend.text.size = 7,multiplot = T, multiplot.layout=rbind(c(1,1,2,3),c(1,1,4,5),c(6,7,8,9)))
   
   
   
   a$'1'
   
}
\author{
Martin Vicencio
}
