% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitGroups.R
\name{splitGroups}
\alias{splitGroups}
\alias{splitGroups}
\alias{balancedSplitGroups}
\title{Split Groups}
\usage{
splitGroups(dataset, groups = unique(dataset$dataset_group), splitter)

balancedSplitGroups(dataset, groups = unique(dataset$dataset_group), splitter)
}
\arguments{
\item{dataset}{dataframe, dataset with groups.}

\item{groups}{character vector, groups to split. It search on \code{dataset$dataset_group}}

\item{splitter}{numeric vector, indicating how many new groups are made form the split, and the portion that takes from old groups. See details.}
}
\value{
Returns a dataset with the groups splitted into new groups.
}
\description{
Takes a dataset with groups, and it split them in new groups made of specific portions of the old ones. 
Balanced version would keep response rates from original groups in each partition that will conform the new ones.
}
\details{
Suposse that you have trhee groups in the dataset \code{old_1,old_2 & old_3} and you want to split the first and third in two new groups 
names \code{new_1 & new_2}, and you want to \code{new_1} take the half of each group and \code{new_2} take a quarter. The usage would be like
\code{splitGroups(dataset,groups = c("old_1","old_3"), splitter = c(new_1=0.5, new_2=0.25))}.
}
\author{
Martin Vicencio
}
