% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlGsub.R
\name{sqlGsub}
\alias{sqlGsub}
\title{Query Matching and Replacement}
\usage{
sqlGsub(query, param=c())
}
\arguments{
\item{query}{character, it contain the query structure and the keywords between @'s.}

\item{param}{character vector, it contain the keywords that would be replaced in query.}
}
\value{
It returns a character to be used in a SQL query.
}
\description{
It takes general queries and returns specific ones by replacing keywords in the first.
}
\details{
param must be indexed by the keyword.
}
\examples{
q="SELECt @col@ FROM @table@ WHERE @cond@"
p=c(col="year",table="FIFAWorldCup",cond="winner LIKE Brazil")
sqlGsub(q,p)
}
\author{
Daniel Fischer
}
