% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlServerConn.R
\name{SQL server connection}
\alias{SQL server connection}
\alias{sqlGetConn}
\alias{sqlClose}
\alias{sqlListTables}
\title{SQL server connection}
\usage{
sqlGetConn(server_name)

sqlClose(server_name)

sqlListTables(server_name)
}
\arguments{
\item{server_name}{character, name of the server to connect, in sqlServer list.}
}
\value{
\code{sqlGetConn()} set the connection by evaluate the expression on List "sqlServers".

\code{sqlListTable()} returns a character vector with names of every table in database.

\code{sqlClose()} disconnect from database.
}
\description{
This functions allow to work with DB in differents servers ending connections after every interaction.
}
\details{
sqlServers is
a list built-in sqlGetConn().
}
\examples{
  sqlGetConn(local) #connect with "local" in the list sqlServers.
  sqlListTable(local) # its empty
  sqlClose(local) ## end connection
}
\author{
Daniel Fischer
}
