% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlWriteTable.R
\name{sqlWriteTable}
\alias{sqlWriteTable}
\title{SQL Write Table}
\usage{
sqlWriteTable(server_name, data, table, ...)
}
\arguments{
\item{server_name}{character, name of the DB server from sqlServers list.}

\item{data}{a data frame or coercible to data frame}

\item{table}{character, name of the table to write in DB.}

\item{...}{other parameters passed on to methods.}
}
\value{
returns booleans as it is specificated for dbWriteTable().
}
\description{
Writes, overwrite or appends a data frame to a database table.
It calls dbWriteTable().
}
\details{
It ends the connection inmediately after getting the results. sqlServers is
a list built-in sqlGetConn().
}
\examples{
sqlWriteTable("local",head(iris),"iris")
sqlWriteTable("local",head(mtcars),"MTCARS")
con=sqlGetConn("local")
dbReadTable(con,"iris")
dbReadTable(con,"MTCARS")

}
\seealso{
"sqlServerConn" and "setSqlServers()" documentation in toolkitEntel and "dbWriteTable()"
from DBI for more details.
}
