% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_char.R
\name{standard_char}
\alias{standard_char}
\title{Standardize Character Column Strings}
\usage{
standard_char(x, remove.non.alphanumeric.char = TRUE, ...)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{remove.non.alphanumeric.char}{If FALSE, all non-alphanumeric characters will be unaltered. If TRUE, all non-alphanumeric characters will be replaced by a specified character. Default: TRUE}
}
\value{
A data fram with standardized character columns.
}
\description{
Standardizes all character column strings by converting the strings to lowercase, removing white space, and removing all non-alphanumeric characters and replacing them with a specified character.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 data("starwars")
 test.df <- head(starwars)
 standard_char(test.df)
 standard_char(test.df, replacement.char = "+")
 }
}
}
\seealso{
\code{\link[dplyr]{mutate_if}},\code{\link[dplyr]{funs}}
}
