% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_names.R
\name{standard_names}
\alias{standard_names}
\title{Standard Column Names}
\usage{
standard_names(x, ...)
}
\arguments{
\item{x}{A data frame or tibble.}
}
\value{
A data fram with standardized column names.
}
\description{
Standardizes column names by converting the strings to lowercase, removing white space, and removing all non-alphanumeric characters and replacing them with a specified character.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 data("iris")
 test.df <- head(iris)
 standard_names(test.df)
 standard_names(test.df, replacement.char = "+")
 }
}
}
