% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simple.R
\name{theme_simple}
\alias{theme_simple}
\title{Personal ggplot theme}
\usage{
theme_simple(
  base_family = "HelveticaNowDisplay Regular",
  base_size = 16,
  strip_text_family = "HelveticaNowDisplay Regular",
  strip_text_size = 16,
  plot_title_family = "HelveticaNowDisplay Bold",
  plot_title_size = 20,
  plot_title_margin = 10,
  subtitle_family = "HelveticaNowDisplay Regular",
  subtitle_size = 16,
  subtitle_margin = 10,
  caption_family = "HelveticaNowDisplay Regular",
  caption_size = 12,
  caption_margin = 10,
  axis_title_family = "HelveticaNowDisplay Regular",
  axis_title_size = 12,
  axis_title_just = "rt",
  dark = FALSE,
  grid = TRUE,
  axis = FALSE,
  ticks = FALSE,
  markdown = TRUE
)
}
\arguments{
\item{base_family}{Base font family}

\item{base_size}{Base font size}

\item{strip_text_family}{Facet label font family}

\item{strip_text_size}{Facet label text size}

\item{plot_title_family}{Plot tilte family}

\item{plot_title_size}{Plot title font size}

\item{plot_title_margin}{Plot title margin}

\item{subtitle_family}{Plot subtitle family}

\item{subtitle_size}{Plot subtitle size}

\item{subtitle_margin}{Plot subtitle margin}

\item{caption_family}{Plot caption family}

\item{caption_size}{Plot caption size}

\item{caption_margin}{Plot caption margin}

\item{axis_title_family}{Axis title font family}

\item{axis_title_size}{Axis title font size}

\item{axis_title_just}{Axis title font justification \code{blmcrt}}

\item{dark}{Dark mode \code{TRUE}, \code{FALSE}}

\item{grid}{Panel grid (\code{TRUE}, \code{FALSE}, or a combination of
\code{X}, \code{x}, \code{Y}, \code{y})}

\item{axis}{Axis \code{TRUE}, \code{FALSE}, \link{\code{xy}}}

\item{ticks}{Ticks \code{TRUE}, \code{FALSE}}

\item{markdown}{Enabled ggtext markdown styling  \code{TRUE}, \code{FALSE}}
}
\description{
A precise, clean and minimal theme for ggplot2. Influenced by @hrbrmstr and @jkaupp.
}
