% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOPALS_fit.R
\name{TOPALS_fit}
\alias{TOPALS_fit}
\title{TOPALS fitting function for mortality schedules (by Carl Schmertmann)}
\usage{
TOPALS_fit(
  N,
  D,
  std,
  age_group_bounds = 0:100,
  knot_positions = c(0, 1, 12, 20, 40, 70),
  penalty_precision = 2,
  max_iter = 50,
  alpha_tol = 5e-05,
  details = FALSE
)
}
\description{
Fits TOPALS parameters to single-year or age-group (D,N)
data by penalized IRLS with analytical derivatives.
}
\details{
\itemize{
\item added fitted log mort, basis, etc to detailed output
\item added ability to fit age-grouped as well as single-yr data
\item changed fitting algorithm from Newton-Raphson
to (penalized) IRLS
age_group_bounds is a (G+1) vector of ages v that define
G closed age groups [v1,v2), [v2,v3)... [vG,vG+1)
For ex. if age_group_bounds = c(0,1,5,10,15,...,90)
then the age groups are [0,1), [1,5), [5,10), ..., [85,90)
A more complete explanation is in
https://github.com/schmert/TOPALS/blob/master/TOPALS_fitting_with_grouped_data.pdf
}
}
