% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toQuantiles.R
\name{toQuantiles}
\alias{toQuantiles}
\title{To Quantiles}
\usage{
toQuantiles(values,n)
}
\arguments{
\item{values}{array of values. The values must be able to rank.}

\item{n}{number of Quantiles to make the assignment. Takes n=100 by default to assign a percentile to every value.}
}
\value{
It returns a List of factors with the percentile of cut of each Quantile.
}
\description{
Takes an array of values and assign Quantiles to each one.
}
\author{
Daniel Fischer
}
