% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainModel.R
\name{trainModel}
\alias{trainModel}
\title{Train Model}
\usage{
trainModel(model,dataset,subset,features=attr(dataset,"features"),...)
}
\arguments{
\item{model}{character that indicates the type of model to be calibrated. Like \code{"randomForest"} or \code{"rpart"}}

\item{dataset}{dataframe with data to calibrate the model.}

\item{subset}{character or character vector, indicate the group(s) of \code{dataset$dataset_group} to be used in calibration.}

\item{features}{features list of dataset}

\item{...}{parameter to be passed on models functions and methods}
}
\value{
It returns a calibration object, the class of this object will depend on the method declarated in \code{model}. And adds
 the features list as an attribute called "features" to it.
}
\description{
It calibrates a model by spicifing the type of model and  giving a dataset with its features list.
}
\details{
This function automatically generates a inner formula using \code{dataset} and \code{features}, so do not pass the formula in \code{...}.
Use \code{...} to define specific parameters for each method like \code{ntree,importance,control,minsplit,maxdepth,} etc.
}
\author{
Martin Vicencio
}
