% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{tricolour_pal}
\alias{tricolour_pal}
\title{Personal palettes}
\format{
An object of class \code{character} of length 3.
}
\usage{
tricolour_pal
}
\description{
Color scales for theme_simple
}
\details{
Semiotic
}
\keyword{datasets}
