% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim}
\usage{
trim(text)
}
\arguments{
\item{text}{character to be trimed.}
}
\value{
Returns a character without spaces at the begin and the end.
}
\description{
Remove the spaces of text, except the spaces between words.
}
\author{
Daniel Fischer
}
