% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueMerge.R
\name{uniqueMerge}
\alias{uniqueMerge}
\title{Unique Merge}
\usage{
uniqueMerge(x,y,by=intersect(names(x),names(y)),
by.x=by,by.y=by,all=T,all.x=all,all.y=all,...)
}
\arguments{
\item{x, y}{data frames, or data tables to be coerced to one.}

\item{by, by.x, by.y}{specifications of the columns used for merging.}

\item{all, all.x, all.y}{logicals, TRUE by default. Adds not matching rows to the output
filling the fields with NAs}

\item{...}{arguments to be passed to or from methods. As sort, suffixes or incomparables.}
}
\description{
Merges data frames or data tables by columns without all possible combination of rows. So if there are
N rows on a table that matches with M rows on the other, this function would only merge the first with
the first, second with second, and go on... Finaly it would be |M-N| not merged rows.
}
\details{
This function calls merge so it admits all the ways to use that function but without the
combination for matching rows.
}
\author{
Daniel Fischer
}
