% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFeatures.R
\name{updateFeatures}
\alias{updateFeatures}
\title{Update Features}
\usage{
updateFeatures(dataset.file,features.file,ds.sep=";",f.sep="\t")
}
\arguments{
\item{dataset.file}{character vector, path of the dataset(s) file.}

\item{features.file}{character, path of the features file, or where it will be wrote.}

\item{ds.sep}{separator used in the trainset.}

\item{f.sep}{separator used in the features file.}
}
\value{
It return nothing, but rewrite the features file with no copie off the old one. If there are variables in the old
features file that any of \code{dataset.file} has not, they will be keeped at the end of features file.
}
\description{
It reads the header of the dataset(s) that will be used, and it updates the features file, adding the new variables as rows filled of
\code{NA\'s & 0}. If there are no features file, it make an empty one from the dataset(s).
}
\author{
Martin Vicencio
}
