% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visionBDA.R
\name{visionBDA}
\alias{visionBDA}
\title{Vision BDA}
\usage{
visionBDA(x_0=0,y_0=0,direction=0,aperture=0,radius=1,steps=0,alpha=0.3,DEG=T)
}
\arguments{
\item{x_0}{numeric, x coordenate of the polygon's center.}

\item{y_0}{numeric, y coordenate of the polygon's center.}

\item{direction}{numeric indicating the orientation of the vision source, it uses the cartography convention where "0" points to north and it increase clockwise.}

\item{aperture}{numeric indicating the aperture of the vision field.}

\item{radius}{numeric, vision radius.}

\item{steps}{integer, number of steps to draw the vision field, the more steps more curvate it gets. If \code{steps=0} then take one step for Degree.}

\item{alpha}{numeric between 0-1 indicating the opacity of the polygon.}

\item{DEG}{boolean, if its \code{FALSE} uses Radians instead of Degrees.}
}
\value{
It returns a list with two elements. \code{\$polygon} is a dataframe with the points of the polygon and \code{\$plot} is a ggplot obect to be add to a ggplot.
}
\description{
Draw and get a orientated vision field
}
\details{
To plot this polygon do something like \code{ggplot()+visionBDA(...)$plot}.
}
\examples{
   p=visionBDA(direction=30,aperture=25,radius=2,alpha=0.4,color="green")
   
   p$polygon
   
   ggplot()+p$plot 
}
\author{
Martin Vicencio
}
