% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_the_hell_is_it.R
\name{where_the_hell_is_it}
\alias{where_the_hell_is_it}
\title{Find string source in a function definition}
\usage{
where_the_hell_is_it(str)
}
\arguments{
\item{str}{Character to search for}
}
\description{
Occasionally I want to track down the source across all
packages of a piece of code. This solution was motivated by
trying to track down use of an unexported function from a package.
This function searches all function definitions in your package library
(both exports and non-exports) for a particular string.
}
\examples{
\dontrun{
 where_the_hell_is_it("yaml_load_file")
}
}
