% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{with_scratch}
\alias{with_scratch}
\title{Execute code inside a scratch directory context.}
\usage{
with_scratch(
  source_directory,
  target_directory,
  modifier = modifier_timed_directory,
  callback_function,
  cleanup_after = FALSE,
  copy_pattern = NULL
)
}
\arguments{
\item{source_directory}{The source directory.}

\item{target_directory}{The target directory.}

\item{modifier}{A target path modifier.}

\item{callback_function}{The function to run inside the scratch directory
context.}

\item{cleanup_after}{Clean up after execution.}

\item{copy_pattern}{Pattern used to copy files from the target_directory back
to the source directory.}
}
\value{
Timed directory with the format \%Y\%m\%d_\%H\%M\%S appended to
target_directory path.
}
\description{
Execute code inside a scratch directory context.
}
