% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D-Splines_fit.R
\name{xpnd}
\alias{xpnd}
\title{functions from MCMCpack to collapse and
re-expand symmetric matrices}
\usage{
xpnd(x, nrow = NULL)
}
\description{
functions from MCMCpack to collapse and
re-expand symmetric matrices
}
