% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_age.R
\name{adjust_age}
\alias{adjust_age}
\title{Adjusting population age structure}
\usage{
adjust_age(
  data,
  location = "Estado de São Paulo",
  year = 2010,
  AgeL = 0,
  AgeH = 75
)
}
\description{
Smooth population counts using Arriaga and Sprague methods
}
