% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Creates pretty formatted tables with few lines of code for both html and pdfs}
\usage{
format_table(
  data = NULL,
  col_bold = FALSE,
  font_size = 15,
  caption = NULL,
  ...
)
}
\arguments{
\item{data}{dataframe, tibble or tableone object with table content}

\item{col_bold}{should first column be printed bold}

\item{font_size}{numeric, font size of table}

\item{caption}{string, if table caption is desired}

\item{...}{more parameters passed on}
}
\value{
returns a pretty table formatted for html or latex
}
\description{
Creates pretty formatted tables with few lines of code for both html and pdfs
}
